(#4) move interfaces to seperate directory
