+++ banking_system/premium_account_manager_factory.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "manager_factory.h"
+#include "interfaces/i_manager_factory.h"
 
 namespace bank {
-    class PremiumAccountManagerFactory: public ManagerFactory {
+    class PremiumAccountManagerFactory: public IManagerFactory {
     public:
         std::unique_ptr<IManager> createManager() override;
     };
